requirements.txt
streamlit
pycryptodome
pyperclip
